requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
django-formtools==2.4
django-otp==1.1.4
django-phonenumber-field==6.4.0
django-two-factor-auth==1.14.0
djangorestframework==3.14.0
phonenumbers==8.13.2
pytz==2022.7
qrcode==7.3.1
sqlparse==0.4.3
